requirements.txt
json-tricks==3.17.3
keras== 2.15.0
librosa==0.10.1
matplotlib==3.7.1
noisereduce==3.0.2
numpy==1.24.2
opencv-python==4.8.0.76
Pillow==9.4.0
scikit-learn==1.4.0
seaborn==0.12.2
streamlit==1.30.0
streamlit-audiorecorder==0.0.4
tensorflow==2.15.0
